+++ packages/ragbits-chat/CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
 ## Unreleased
+- Add authentication handling in the UI (#763)
+- Add backend authentication (#761)
+
+- Autogenerate typescript types based on backend typing (#727)
+- Add ability to customize favicon and page title (#766)
 
 - Autogenerate typescript types based on backend typing (#727)
 - Add ability to customize favicon and page title (#766)